# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API DreamCar Reserva de vehiculos",
        "description": "Esta API nos permitira ver y manipular informacion a las reservas de los vehiculos.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/rest/reservas/show/{id}": {
            "get": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Obtener una reserva por id",
                "description": "Renorna los detalles de una reserva",
                "operationId": "e18c6cbf036768051a8f747f57a4c93d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reserva",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la reserva con id seleccionado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "email": {
                                            "type": "string",
                                            "maxLength": 50,
                                            "example": "Email de la reserva",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rest/reservas/reserva/{email_cliente}": {
            "get": {
                "tags": [
                    "Reserva"
                ],
                "summary": "Buscar reserva por email del cliente",
                "description": "Retorna una reserva",
                "operationId": "9ae85ebcee4289fe54cbe84b9a195b6a",
                "parameters": [
                    {
                        "name": "email_cliente",
                        "in": "path",
                        "description": "email de la reserva a buscar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "email": {
                                                "type": "string",
                                                "maxLength": 50,
                                                "example": "Reserva por email",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontro la reserva",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rest/reservas/reserva/{id}": {
            "get": {
                "tags": [
                    "Reserva"
                ],
                "summary": "Mostrar los detalles de una reserva",
                "description": "Retorna los detalles de una reserva",
                "operationId": "d707180d666f2baaaa7fc3eb3dd5f9d1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id de la reserva a buscar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la reserva encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "data": {
                                                "type": "string",
                                                "maxLength": 1000,
                                                "example": "Detalles de la reserva",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontro la Reserva",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rest/reservas/crearReserva": {
            "post": {
                "tags": [
                    "Reserva"
                ],
                "summary": "Crear reserva",
                "description": "Crea una nueva reserva en base a un email de un cliente",
                "operationId": "c22c32bb354512cd7a3a6007dcfbe7a4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "test@gmail.com"
                                    },
                                    "fecha_inicio": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-05-18"
                                    },
                                    "fecha_final": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-05-18"
                                    },
                                    "vehiculos": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reserva creada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensaje": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error al crear la reserva"
                    }
                }
            }
        },
...
    },
...
}